Task-46549: refactoring of news body
