+++ webapp/src/main/webapp/news-details/initComponents.js
@@ -4,8 +4,9 @@ import ExoNewsDetailsActionMenu from './components/ExoNewsDetailsActionMenu.vue'
 import ExoNewsPin from './components/ExoNewsPin.vue';
 import ExoNewsDetailsActivity from './components/ExoNewsDetailsActivity.vue';
 import ExoNewsDetailsToolBar from './components/ExoNewsDetailsToolBar.vue';
-import ExoNewsDetailsToolBarMobile from './components/ExoNewsDetailsToolBarMobile.vue';
-import ExoNewsDetailsBodyMobile from './components/ExoNewsDetailsBodyMobile.vue';
+import ExoNewsDetailsToolBarMobile from './components/mobile/ExoNewsDetailsToolBarMobile.vue';
+import ExoNewsDetailsBodyMobile from './components/mobile/ExoNewsDetailsBodyMobile.vue';
+import ExoNewsDetailsBody from './components/ExoNewsDetailsBody.vue';
 
 const components = {
   'exo-news-archive': ExoNewsArchive,
@@ -16,6 +17,7 @@ const components = {
   'exo-news-details-toolbar': ExoNewsDetailsToolBar,
   'exo-news-details-toolbar-mobile': ExoNewsDetailsToolBarMobile,
   'exo-news-details-body-mobile': ExoNewsDetailsBodyMobile,
+  'exo-news-details-body': ExoNewsDetailsBody,
 };
 
 for (const key in components) {
