+++ webapp/src/main/webapp/skin/less/news.less
@@ -13,11 +13,6 @@
 
 .VuetifyApp {
   #newsDetail {
-    .newsDetailsMobileBody {
-      .newsBodyHeader {
-        margin-top: 35%!important;
-      }
-    }
     padding: 10px 8%;
     .articleNotFound {
       text-align: center;
@@ -2882,6 +2877,10 @@
   margin-right: 11% ~'; /** orientation=rt */ ';
   margin-left: 10% ~'; /** orientation=rt */ ';
   margin-top: 21px;
+}
+
+.news-details-toolbar {
+  position: inherit!important;
 }
 
 /*====================== Responsive ==========================*/
